#38 [REFACTOR] 탭 레이아웃 요구사항에 맞게 수정
