+++ app/src/main/java/org/woowatechcamp/githubapplication/presentation/home/MainActivity.kt
@@ -1,13 +1,13 @@
-package org.woowatechcamp.githubapplication.presentation
+package org.woowatechcamp.githubapplication.presentation.home
 
 import android.content.Intent
+import android.graphics.Bitmap
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
-import android.view.MotionEvent
 import androidx.activity.viewModels
+import androidx.core.graphics.drawable.RoundedBitmapDrawable
 import androidx.core.graphics.drawable.RoundedBitmapDrawableFactory
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayoutMediator
@@ -15,8 +15,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.woowatechcamp.githubapplication.R
 import org.woowatechcamp.githubapplication.databinding.ActivityMainBinding
 import org.woowatechcamp.githubapplication.presentation.adapter.ViewpagerAdapter
-import org.woowatechcamp.githubapplication.presentation.issue.IssueFragment
-import org.woowatechcamp.githubapplication.presentation.notifications.NotificationsFragment
+import org.woowatechcamp.githubapplication.presentation.home.issue.IssueFragment
+import org.woowatechcamp.githubapplication.presentation.home.notifications.NotificationsFragment
 import org.woowatechcamp.githubapplication.presentation.profile.ProfileActivity
 import org.woowatechcamp.githubapplication.util.showSnackBar
 import kotlin.math.max
@@ -50,7 +50,7 @@ class MainActivity : AppCompatActivity() {
         binding.vpMain.apply {
             adapter = viewpagerAdapter
             orientation = ViewPager2.ORIENTATION_HORIZONTAL
-            isUserInputEnabled = true // 사용자가 직접 swipe 할 수 있도록 하는 것
+            isUserInputEnabled = false // 사용자가 직접 swipe 할 수 있도록 하는 것
         }
         TabLayoutMediator(binding.tlMain, binding.vpMain) { tab, position ->
             tab.text = viewpagerAdapter.getTitle(position)
@@ -62,15 +62,24 @@ class MainActivity : AppCompatActivity() {
             showSnackBar(binding.root, it, this)
         }
         mViewModel.userProfile.observe(this) {
-            val round = RoundedBitmapDrawableFactory.create(resources, it)
-            round.cornerRadius = max(it.width, it.height) / 2f
-            menu.getItem(1).icon = round
+            if (::menu.isInitialized) {
+                menu.getItem(1).icon = getRoundDrawable(it)
+            }
         }
+    }
+
+    private fun getRoundDrawable(bitmap : Bitmap) : RoundedBitmapDrawable {
+        val round = RoundedBitmapDrawableFactory.create(resources, bitmap)
+        round.cornerRadius = max(bitmap.width, bitmap.height) / 2f
+        return round
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         this.menu = menu
         menuInflater.inflate(R.menu.menu_main, menu)
+        mViewModel.userProfile.value?.let {
+            menu.getItem(1).icon = getRoundDrawable(it)
+        }
         return true
     }
 
@@ -80,9 +89,13 @@ class MainActivity : AppCompatActivity() {
 
             }
             R.id.menu_main_profile -> {
-                val intent = Intent(this, ProfileActivity::class.java)
-                intent.putExtra("profile_item", mViewModel.userInfo.value)
-                startActivity(intent)
+                mViewModel.userInfo.value?.let {
+                    val intent = Intent(this, ProfileActivity::class.java)
+                    intent.putExtra("profile_item", it.getModel())
+                    startActivity(intent)
+                    return true
+                }
+                showSnackBar(binding.root, "데이터를 불러오는 데 실패했습니다.", this)
             }
             else -> return true
         }
