+++ app/src/main/java/org/woowatechcamp/githubapplication/presentation/home/notifications/NotificationsFragment.kt
@@ -1,24 +1,23 @@
-package org.woowatechcamp.githubapplication.presentation.notifications
+package org.woowatechcamp.githubapplication.presentation.home.notifications
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.ItemTouchHelper
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import dagger.hilt.android.AndroidEntryPoint
 import org.woowatechcamp.githubapplication.R
 import org.woowatechcamp.githubapplication.data.notifications.model.NotiResponseItem
 import org.woowatechcamp.githubapplication.databinding.FragmentNotificationsBinding
-import org.woowatechcamp.githubapplication.presentation.notifications.adapter.NotiAdapter
-import org.woowatechcamp.githubapplication.presentation.view_util.CustomItemDivider
-import org.woowatechcamp.githubapplication.presentation.view_util.SwipeCallback
-import org.woowatechcamp.githubapplication.presentation.view_util.SwipeListener
+import org.woowatechcamp.githubapplication.presentation.home.notifications.adapter.NotiAdapter
+import org.woowatechcamp.githubapplication.util.ItemDecorationUtil
+import org.woowatechcamp.githubapplication.util.SwipeCallback
+import org.woowatechcamp.githubapplication.util.SwipeListener
+import org.woowatechcamp.githubapplication.util.ext.getDeliNumber
 import org.woowatechcamp.githubapplication.util.showSnackBar
 
 @AndroidEntryPoint
@@ -46,7 +45,6 @@ class NotificationsFragment : Fragment(), SwipeListener, SwipeRefreshLayout.OnRe
         observeData()
 
         binding.swipeNoti.setOnRefreshListener(this)
-
         mVieModel.getNoti()
     }
 
@@ -54,10 +52,7 @@ class NotificationsFragment : Fragment(), SwipeListener, SwipeRefreshLayout.OnRe
         notiAdapter = NotiAdapter()
         binding.rvNoti.apply {
             adapter = notiAdapter
-            addItemDecoration(CustomItemDivider(1f, requireActivity().getColor(R.color.navy)))
-            layoutManager = LinearLayoutManager(
-                requireActivity(), LinearLayoutManager.VERTICAL, false
-            )
+            addItemDecoration(ItemDecorationUtil.ItemDividerDecoration(1f, 0f, requireActivity().getColor(R.color.navy)))
         }
         swipeCallback = SwipeCallback(0, ItemTouchHelper.LEFT, requireActivity())
         swipeCallback.setListener(this)
@@ -65,8 +60,9 @@ class NotificationsFragment : Fragment(), SwipeListener, SwipeRefreshLayout.OnRe
     }
 
     private fun observeData() {
+        // 현재 아예 아무 응답도 안 와서 성공해도 오류 메시지가 뜸
         mVieModel.markMessage.observe(viewLifecycleOwner) {
-            showSnackBar(binding.root, it, requireActivity())
+//            showSnackBar(binding.root, it, requireActivity())
         }
         mVieModel.notiList.observe(viewLifecycleOwner) {
             notiAdapter.submitList(it)
@@ -86,11 +82,9 @@ class NotificationsFragment : Fragment(), SwipeListener, SwipeRefreshLayout.OnRe
         list.removeAt(position)
         notiAdapter.submitList(list.toList())
         // thread ID 값을 가져와서 알림 읽음 설정한다.
-        val urls = item.url.split("threads/")
-        if (urls.size > 1) {
-            val num = urls[1].toLong()
-            Log.d("HELLO", "thread num = ${num}")
-            mVieModel.markNoti(num)
+        val threadNum = item.url.getDeliNumber("threads/")
+        if (threadNum > 0) {
+            mVieModel.markNoti(threadNum)
         }
     }
