+++ app/src/main/res/drawable/tab_main_selected.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:right="8dp" android:left="8dp">
+        <shape>
+            <corners android:radius="14dp"/>
+            <solid  android:color="@color/primary"/>
+        </shape>
+    </item>
+</layer-list>
